Factor all the logic into the board component, pending major refactor
