+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import "./App.css";
-import Board from "./components/Board";
-import GameBoard from "./board";
+// import Board from "./components/Board";
+// import GameBoard from "./board";
 import GameLoop from "./GameLoop";
 
 function App() {
