--- src/ship.js
@@ -1,83 +1,0 @@
-export default class Ship {
-	constructor(length, x, y, rotation) {
-		this.length = length || this._getRandomLength(5);
-		this.hits = 0;
-		this.x = typeof x === "number" ? x : null;
-		this.y = typeof y === "number" ? y : null;
-		this.rotation = typeof rotation === "number" ? rotation : null;
-		this._coordList =
-			this.x !== null && this.x !== null && this.x !== null
-				? this.getShipCoords(
-						this.length,
-						[
-							this._testPositiveIntOrZero(x),
-							this._testPositiveIntOrZero(y),
-						],
-						this._testPositiveIntOrZero(rotation)
-				  )
-				: null;
-	}
-
-	getShipCoords(arglen, [argx, argy], argrotation) {
-		if (this._coordList) return this._coordList;
-		const [x, y] = [argx, argy] || [this.x, this.y];
-		const rotation = argrotation || this.rotation;
-		const len = arglen || this.length;
-		const coordList = Ship.getShipCoords(len, [x, y], rotation);
-		this._coordList = coordList;
-		return coordList;
-	}
-
-	static getShipCoords(len, [x, y], rotation) {
-		const coordList = [];
-
-		const isRotEven = rotation % 2 === 0;
-
-		for (let i = 0; i < len; i++) {
-			const newX = isRotEven ? x : rotation < 2 ? x + i : x - i;
-
-			const newY = !isRotEven ? y : rotation < 2 ? y - i : y + i;
-			coordList.push([newX, newY]);
-		}
-		return coordList;
-	}
-
-	isSunk() {
-		return this.hits >= this.length;
-	}
-
-	hit() {
-		this.hits += 1;
-	}
-
-	set length(len) {
-		this._length = this._testPositiveInt(len);
-	}
-	get length() {
-		return this._length;
-	}
-
-	_testPositiveIntOrZero(input) {
-		const num = parseInt(input);
-		if (num !== Number(num)) {
-			throw new Error(
-				"Wrong type or format (" + num + "), expected a positive int"
-			);
-		}
-		return num;
-	}
-
-	_testPositiveInt(input) {
-		const num = parseInt(input);
-		this._testPositiveIntOrZero(num);
-		if (num < 1) {
-			throw new Error(
-				"Wrong type or format (" + num + "), expected a positive int"
-			);
-		}
-		return num;
-	}
-	_getRandomLength(max) {
-		return Math.floor(Math.random() * max) + 1;
-	}
-}
